+++ src/main/java/roomescape/interface_adapter/ReservationTime/ReservationTimeRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.interface_adapter;
+package roomescape.interface_adapter.ReservationTime;
 
 import java.time.LocalTime;
