+++ src/main/java/roomescape/interface_adapter/ReservationTime/ReservationTimeResponseDto.java
@@ -1,7 +1,7 @@
-package roomescape.interface_adapter;
+package roomescape.interface_adapter.ReservationTime;
 
 import java.time.LocalTime;
-import roomescape.usecase.ReservationTimeOutput;
+import roomescape.usecase.ReservationTime.ReservationTimeOutput;
 
 public record ReservationTimeResponseDto(Long id, LocalTime start_at) {
     public static ReservationTimeResponseDto from(final ReservationTimeOutput reservationTimeOutputModel) {
