+++ src/main/java/roomescape/usecase/ReservationTime/DeleteReservationTimeUsecase.java
@@ -1,4 +1,4 @@
-package roomescape.usecase;
+package roomescape.usecase.ReservationTime;
 
 public interface DeleteReservationTimeUsecase {
     void deleteReservationTime(long id);
